Refactor dev UI package to tvhome and update imports
