+++ app/release/output-metadata.json
@@ -11,8 +11,8 @@
       "type": "SINGLE",
       "filters": [],
       "attributes": [],
-      "versionCode": 12,
-      "versionName": "1.1.10",
+      "versionCode": 13,
+      "versionName": "1.1.11",
       "outputFile": "app-release.apk"
     }
   ],
