+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/screens/DebugScreen.kt
@@ -42,7 +42,6 @@ import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.automirrored.filled.DirectionsRun
 import androidx.compose.material.icons.sharp.Info
@@ -51,22 +50,14 @@ import androidx.compose.runtime.*
 import com.skylake.skytv.jgorunner.data.SkySharedPref
 import com.skylake.skytv.jgorunner.ui.components.ButtonContent
 
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.automirrored.filled.PlaylistPlay
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.font.FontWeight
-import androidx.core.content.ContextCompat.startActivity
 import com.skylake.skytv.jgorunner.activities.CastActivity
-import com.skylake.skytv.jgorunner.activities.ChannelInfo
-import com.skylake.skytv.jgorunner.activities.ExoplayerActivity
-import com.skylake.skytv.jgorunner.activities.ExoplayerActivityPass
 import com.skylake.skytv.jgorunner.services.player.LandingPage
-import com.skylake.skytv.jgorunner.ui.dev.changeIconTOFirst
-import com.skylake.skytv.jgorunner.ui.dev.changeIconToSecond
-import com.skylake.skytv.jgorunner.ui.dev.switchIcon
+import com.skylake.skytv.jgorunner.ui.tvhome.changeIconTOFirst
+import com.skylake.skytv.jgorunner.ui.tvhome.changeIconToSecond
 
 @Composable
 fun DebugScreen(context: Context, onNavigate: (String) -> Unit) {
