+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/screens/ZoneScreen.kt
@@ -57,11 +57,11 @@ import androidx.compose.ui.unit.sp
 import com.skylake.skytv.jgorunner.R
 import com.skylake.skytv.jgorunner.data.SkySharedPref
 import com.skylake.skytv.jgorunner.ui.components.ModeSelectionDialog2
-import com.skylake.skytv.jgorunner.ui.dev.Main_Layout
-import com.skylake.skytv.jgorunner.ui.dev.Recent_Layout
-import com.skylake.skytv.jgorunner.ui.dev.Main_Layout_3rd
-import com.skylake.skytv.jgorunner.ui.dev.depreciated.SearchTabLayout
-import com.skylake.skytv.jgorunner.ui.dev.depreciated.SearchTabLayoutTV
+import com.skylake.skytv.jgorunner.ui.tvhome.Main_Layout
+import com.skylake.skytv.jgorunner.ui.tvhome.Recent_Layout
+import com.skylake.skytv.jgorunner.ui.tvhome.Main_Layout_3rd
+import com.skylake.skytv.jgorunner.ui.tvhome.depreciated.SearchTabLayout
+import com.skylake.skytv.jgorunner.ui.tvhome.depreciated.SearchTabLayoutTV
 import kotlinx.coroutines.launch
 import kotlin.random.Random
