+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/tvhome/ChannelGrids.kt
@@ -1,4 +1,4 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.tvhome
 
 import android.content.Context
 import android.content.Intent
