+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/tvhome/ChannelResponse.kt
@@ -1,4 +1,4 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.tvhome
 
 import androidx.annotation.Keep
