+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/tvhome/Main_Layout_3rd.kt
@@ -1,4 +1,4 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.tvhome
 
 import android.content.Context
 import android.content.Intent
